split speaker recognition type into 2 types; speakerVerification and speackerIdentefication to match js source code
